Add file logging hook with rotation support
